I think the problem is that there is some data left at the end of the file.  I'm not sure if I'm calculating the length incorrectly or if it's just adding junk or what.

+++ Common.java
@@ -51,9 +51,9 @@ public class Common
       return result;
    }
 
-   public static long getNewFormatLength(int[] bytes)
+   public static int getNewFormatLength(int[] bytes)
    {
-      long result = 0;
+      int result = 0;
       if(bytes.length == 1)
       {
          result = bytes[0];
